Stop further processing when autoprogressing into the next tree (#6)
Previously, when the CLI would automatically progress into the next tree
element via recursion, when it returned from completing the first
download, it would fall into the lower multiple choice block to ask the
user what to do next. After answering enough questions, the CLI would
naturally exit, but we can easily and reasonably skip asking these
questions in the first place as per the intent of the original feature.